mess with the renderer flags to increase performance - don't preserve the buffer anymore and clear between frames;
gives a FPS kick to underpowered devices like firestick and chromecast,
going from ~20fps to 60fps on chromecast and 6fps to ~50fps on firestick

this breaks the R export which generates a smart smear type of thingy. could put it back via a dedicated renderer; will look into that later. the R
is not necessarily core to loopkit but would be nice to have back again
at one point